refactor: extract BranchSwitcher component and useBranchSwitcher hook
